WEB-4454 revert moving hook into own folder
