+++ app/pages/clinicworkspace/DeviceIssues/DeviceIssues.js
@@ -9,7 +9,7 @@ import { DIABETES_TYPES } from '../../../core/constants';
 import { RTKQueryApi } from '../../../redux/api/baseApi';
 import { TagList } from '../../../components/elements/Tag';
 
-import useClinicPatientsFilters, { defaultFilterState } from '../hooks/useClinicPatientsFilters';
+import useClinicPatientsFilters, { defaultFilterState } from '../useClinicPatientsFilters';
 import ActiveFilterCount from '../components/ActiveFilterCount';
 import FilterByTags from '../components/FilterByTags';
 import FilterByCategory, { CATEGORY_TAB } from './FilterByCategory';
